requirements.txt
asgiref==3.7.2
crispy-bootstrap5==2024.2
distlib==0.3.8
Django==5.0.2
django-admin-interface==0.28.5
django-colorfield==0.11.0
django-crispy-forms==2.1
django-registration-redux==2.13
django-tinymce==3.7.1
django-versatileimagefield==3.1
djangorestframework==3.15.1
filelock==3.13.1
mysqlclient==2.2.1
pillow==10.2.0
platformdirs==4.1.0
python-decouple==3.8
python-magic==0.4.27
python-magic-bin==0.4.14
python-slugify==8.0.4
sqlparse==0.4.4
text-unidecode==1.3
tzdata==2024.1
virtualenv==20.25.0
